requirements.txt
pytesseract
pillow
openai
pydantic